+++ vkworkspace/__meta__.py
@@ -1,2 +1,2 @@
-__version__ = "1.8.9"
+__version__ = "1.9.0"
 __api_version__ = "v1"