+++ vkworkspace/filters/__init__.py
@@ -2,20 +2,44 @@ from .base import BaseFilter
 from .callback_data import CallbackData, CallbackDataFactory
 from .chat_type import ChatTypeFilter
 from .command import Command, CommandObject
-from .message_parts import ForwardFilter, RegexpPartsFilter, ReplyFilter
+from .message_parts import (
+    AudioFilter,
+    CallbackDataRegexpFilter,
+    FileFilter,
+    ForwardFilter,
+    ImageFilter,
+    MentionFilter,
+    RegexpPartsFilter,
+    ReplyFilter,
+    SenderFilter,
+    StickerFilter,
+    URLFilter,
+    VideoFilter,
+    VoiceFilter,
+)
 from .regexp import RegexpFilter
 from .state import StateFilter
 
 __all__ = [
+    "AudioFilter",
     "BaseFilter",
     "CallbackData",
     "CallbackDataFactory",
+    "CallbackDataRegexpFilter",
     "ChatTypeFilter",
     "Command",
     "CommandObject",
+    "FileFilter",
     "ForwardFilter",
+    "ImageFilter",
+    "MentionFilter",
     "RegexpFilter",
     "RegexpPartsFilter",
     "ReplyFilter",
+    "SenderFilter",
     "StateFilter",
+    "StickerFilter",
+    "URLFilter",
+    "VideoFilter",
+    "VoiceFilter",
 ]